requirements.txt
streamlit
openai
streamlit_pdf_viewer
wordcloud
pdfplumber
reportlab
pillow
pymupdf
scikit-learn
plotly